fix to POSIX intermittent read/write bug (25-Mar-99)
This patch applies to all POSIX platforms.

There is a bug in the procedures FilePosix.IntermittentRead and
FilePosix.IntermittentWrite that sometimes causes the wrong error number to be
reported. To fix the bug, check out the m3/libm3 package, and edit the file
m3/libm3/src/os/POSIX/FilePosix.m3. In the conditional case of "status = -1"
around lines 237 (in IntermittentRead) and 280 (in IntermittentWrite), change
the line:

  OSErrorPosix.Raise()

to instead read:

  OSErrorPosix.Raise0(errno)

Then rebuild and ship the m3/libm3 package.

Thanks to John Polstra for noticing the bug and supplying the fix. This bug
has been fixed in all releases of SRC Modula-3 after 3.6.